feat(web): extract skills stats helpers into skills-stats-lib
Move skill report assembly into skills-stats-lib with expanded Vitest coverage.

Preserve the public @/lib/skills-stats import surface.

Closes #4305
