+++ tests/skills-stats-lib.test.ts
@@ -1,13 +1,14 @@
 import { describe, expect, it } from "vitest";
 
-import { buildSkillsReport } from "../apps/web/src/lib/skills-stats";
+import { buildSkillsReport } from "@/lib/skills-stats-lib";
+import { buildSkillsReport as buildSkillsReportFromWrapper } from "@/lib/skills-stats";
 import {
   buildReportDataset,
   REPORT_PATHS,
   tagDistribution,
-} from "../apps/web/src/lib/data-reports";
-import { ENTRIES, REGISTRY_GENERATED_AT } from "../apps/web/src/data/entries";
-import type { Entry } from "../apps/web/src/types/registry";
+} from "@/lib/data-reports";
+import { ENTRIES, REGISTRY_GENERATED_AT } from "@/data/entries";
+import type { Entry } from "@/types/registry";
 
 function skill(partial: Partial<Entry>): Entry {
   return {
@@ -19,7 +20,7 @@ function skill(partial: Partial<Entry>): Entry {
   } as Entry;
 }
 
-describe("tagDistribution", () => {
+describe("skills-stats-lib tagDistribution", () => {
   it("ranks tags by frequency and honours the exclude set", () => {
     const entries = [
       { tags: ["Testing", "skills"] },
@@ -35,7 +36,7 @@ describe("tagDistribution", () => {
   });
 });
 
-describe("buildSkillsReport (deterministic)", () => {
+describe("skills-stats-lib buildSkillsReport (deterministic)", () => {
   const sample = [
     skill({
       skillType: "capability-pack",
@@ -57,9 +58,11 @@ describe("buildSkillsReport (deterministic)", () => {
     const b = buildSkillsReport(sample, "2026-06-20");
     expect(a).toEqual(b);
     expect(a.slug).toBe("/state-of-agent-skills");
+    expect(a.exportSlug).toBe("agent-skills");
     expect(a.total).toBe(2);
     expect(a.stats.find((s) => s.key === "validated")?.value).toBe(50);
     expect(a.stats.find((s) => s.key === "packs")?.value).toBe(50);
+    expect(a.stats.find((s) => s.key === "packaged")?.value).toBe(50);
   });
 
   it("drops degenerate single-bucket dimensions", () => {
@@ -71,12 +74,75 @@ describe("buildSkillsReport (deterministic)", () => {
     for (const dimension of model.dimensions) {
       expect(dimension.rows.length).toBeGreaterThan(1);
     }
-    // skill-type is uniform here -> must be dropped
     expect(model.dimensions.some((d) => d.key === "skill-type")).toBe(false);
+  });
+
+  it("excludes mechanism tags from the use-case distribution", () => {
+    const model = buildSkillsReport(
+      [
+        skill({ tags: ["skills", "agent-skills", "testing"] }),
+        skill({ tags: ["Skill", "CI", "testing"] }),
+      ],
+      "2026-06-20",
+    );
+    const useCases = model.dimensions.find((d) => d.key === "use-cases");
+    expect(useCases?.rows.map((row) => row.label)).toEqual(["testing", "ci"]);
+  });
+
+  it("reports packaging, maturity, and verification dimensions for diverse samples", () => {
+    const model = buildSkillsReport(
+      [
+        skill({
+          skillType: "capability-pack",
+          skillLevel: "foundational",
+          verificationStatus: "draft",
+          packageVerified: true,
+          tags: ["testing"],
+        }),
+        skill({
+          skillType: "general",
+          skillLevel: "expert",
+          verificationStatus: "production",
+          packageVerified: false,
+          tags: ["docs"],
+        }),
+      ],
+      "2026-06-20",
+    );
+    const keys = model.dimensions.map((dimension) => dimension.key);
+    expect(keys).toEqual(
+      expect.arrayContaining([
+        "skill-type",
+        "maturity",
+        "verification",
+        "packaging",
+      ]),
+    );
+    const packaging = model.dimensions.find((d) => d.key === "packaging");
+    expect(packaging?.rows).toEqual([
+      { label: "Verified package", count: 1, pct: 50 },
+      { label: "Source only", count: 1, pct: 50 },
+    ]);
+  });
+
+  it("returns an empty report when the registry has no skills", () => {
+    const model = buildSkillsReport(
+      [skill({ category: "agents", tags: ["testing"] }) as Entry],
+      "2026-06-20",
+    );
+    expect(model.total).toBe(0);
+    expect(model.dimensions).toEqual([]);
+    expect(model.stats.find((s) => s.key === "total")?.value).toBe(0);
+  });
+
+  it("keeps the public wrapper re-export aligned with the lib module", () => {
+    const fromLib = buildSkillsReport(sample, "2026-06-20");
+    const fromWrapper = buildSkillsReportFromWrapper(sample, "2026-06-20");
+    expect(fromWrapper).toEqual(fromLib);
   });
 });
 
-describe("report Dataset JSON-LD", () => {
+describe("skills-stats-lib report Dataset JSON-LD", () => {
   it("measures every stat and dimension", () => {
     const model = buildSkillsReport(
       [
@@ -87,21 +153,23 @@ describe("report Dataset JSON-LD", () => {
     );
     const ds = buildReportDataset(model) as Record<string, unknown>;
     expect(ds["@type"]).toBe("Dataset");
+    expect(ds.dateModified).toBe("2026-06-20");
     const measured = ds.variableMeasured as string[];
     for (const stat of model.stats) expect(measured).toContain(stat.label);
     for (const dimension of model.dimensions)
       expect(measured).toContain(dimension.title);
+    expect(new Set(measured).size).toBe(measured.length);
   });
 });
 
-describe("sitemap manifest", () => {
+describe("skills-stats-lib sitemap manifest", () => {
   it("lists the new report so it gets indexed", () => {
     expect(REPORT_PATHS).toContain("/state-of-agent-skills");
     expect(new Set(REPORT_PATHS).size).toBe(REPORT_PATHS.length);
   });
 });
 
-describe("real registry data", () => {
+describe("skills-stats-lib real registry data", () => {
   const asOf = String(REGISTRY_GENERATED_AT).slice(0, 10);
   const model = buildSkillsReport(ENTRIES, asOf);
 
@@ -111,12 +179,5 @@ describe("real registry data", () => {
     for (const dimension of model.dimensions) {
       expect(dimension.rows.length).toBeGreaterThan(1);
     }
-    // eslint-disable-next-line no-console
-    console.log(
-      "skills report dimensions:",
-      model.dimensions.map((d) => `${d.key}(${d.rows.length})`).join(", "),
-      "| stats:",
-      model.stats.map((s) => `${s.key}=${s.value}`).join(", "),
-    );
   });
 });
